# tpodashboard/part1/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepVista API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/drives/eligibility-check": {
      "post": {
        "tags": [
          "drives"
        ],
        "summary": "Check Eligibility",
        "description": "Live eligibility preview while a TPO is drafting a drive \u2014 no drive needs to exist yet.",
        "operationId": "check_eligibility_drives_eligibility_check_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EligibilityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drives": {
      "get": {
        "tags": [
          "drives"
        ],
        "summary": "List Drives",
        "operationId": "list_drives_drives_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DriveStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "drives"
        ],
        "summary": "Create Drive",
        "operationId": "create_drive_drives_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriveCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drives/{drive_id}": {
      "get": {
        "tags": [
          "drives"
        ],
        "summary": "Get Drive",
        "operationId": "get_drive_drives__drive_id__get",
        "parameters": [
          {
            "name": "drive_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drive Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "drives"
        ],
        "summary": "Update Drive",
        "operationId": "update_drive_drives__drive_id__patch",
        "parameters": [
          {
            "name": "drive_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drive Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriveUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drives/{drive_id}/eligible-students": {
      "get": {
        "tags": [
          "drives"
        ],
        "summary": "Get Eligible Students",
        "operationId": "get_eligible_students_drives__drive_id__eligible_students_get",
        "parameters": [
          {
            "name": "drive_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drive Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Eligible Students Drives  Drive Id  Eligible Students Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drives/{drive_id}/rounds": {
      "post": {
        "tags": [
          "drives"
        ],
        "summary": "Add Round",
        "operationId": "add_round_drives__drive_id__rounds_post",
        "parameters": [
          {
            "name": "drive_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drive Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewRoundIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRoundOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "drives"
        ],
        "summary": "List Rounds",
        "operationId": "list_rounds_drives__drive_id__rounds_get",
        "parameters": [
          {
            "name": "drive_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drive Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewRoundOut"
                  },
                  "title": "Response List Rounds Drives  Drive Id  Rounds Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriveCreate": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "season_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type",
            "default": "FULL_TIME"
          },
          "ctc_lpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Lpa"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "min_cgpa": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Min Cgpa",
            "default": 0
          },
          "max_backlogs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Backlogs",
            "default": 99
          },
          "eligible_department_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligible Department Ids"
          },
          "application_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Deadline"
          },
          "drive_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Date"
          },
          "status": {
            "$ref": "#/components/schemas/DriveStatus",
            "default": "DRAFT"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "role"
        ],
        "title": "DriveCreate"
      },
      "DriveDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "ctc_lpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Lpa"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "status": {
            "$ref": "#/components/schemas/DriveStatus"
          },
          "min_cgpa": {
            "type": "number",
            "title": "Min Cgpa"
          },
          "max_backlogs": {
            "type": "integer",
            "title": "Max Backlogs"
          },
          "eligible_department_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligible Department Ids"
          },
          "application_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Deadline"
          },
          "drive_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "role",
          "job_type",
          "ctc_lpa",
          "location",
          "status",
          "min_cgpa",
          "max_backlogs",
          "eligible_department_ids",
          "application_deadline",
          "drive_date"
        ],
        "title": "DriveDetail"
      },
...
      "DriveListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "ctc_lpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Lpa"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "status": {
            "$ref": "#/components/schemas/DriveStatus"
          },
          "application_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Deadline"
          },
          "applicant_count": {
            "type": "integer",
            "title": "Applicant Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "role",
          "ctc_lpa",
          "location",
          "status",
          "application_deadline"
        ],
        "title": "DriveListItem"
      },
      "DriveListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DriveListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "DriveListResponse"
      },
      "DriveStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "UPCOMING",
          "ACTIVE",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "DriveStatus"
      },
      "DriveUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "ctc_lpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Lpa"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "min_cgpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Cgpa"
          },
          "max_backlogs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Backlogs"
          },
          "eligible_department_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligible Department Ids"
          },
          "application_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Deadline"
          },
          "drive_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Date"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriveStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "DriveUpdate"
      },
...
      "EligibilityRequest": {
        "properties": {
          "min_cgpa": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Min Cgpa"
          },
          "max_backlogs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Backlogs"
          },
          "department_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Ids"
          }
        },
        "type": "object",
        "required": [
          "min_cgpa",
          "max_backlogs"
        ],
        "title": "EligibilityRequest"
      },
      "EligibilityResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "eligible": {
            "type": "integer",
            "title": "Eligible"
          },
          "not_eligible": {
            "type": "integer",
            "title": "Not Eligible"
          },
          "cgpa_fail": {
            "type": "integer",
            "title": "Cgpa Fail"
          },
          "backlog_fail": {
            "type": "integer",
            "title": "Backlog Fail"
          },
          "dept_fail": {
            "type": "integer",
            "title": "Dept Fail"
          }
        },
        "type": "object",
        "required": [
          "total",
          "eligible",
          "not_eligible",
          "cgpa_fail",
          "backlog_fail",
          "dept_fail"
        ],
        "title": "EligibilityResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewRoundIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sequence_order": {
            "type": "integer",
            "title": "Sequence Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "InterviewRoundIn"
      },
      "InterviewRoundOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sequence_order": {
            "type": "integer",
            "title": "Sequence Order",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "drive_id": {
            "type": "string",
            "format": "uuid",
            "title": "Drive Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "drive_id"
        ],
        "title": "InterviewRoundOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}